Added suggestions and remove planner for nodejs
