+++ internal/pkgsuggest/pkgsuggest.go
@@ -3,7 +3,9 @@ package pkgsuggest
 import (
 	"go.jetpack.io/devbox/internal/pkgsuggest/suggestors"
 	"go.jetpack.io/devbox/internal/pkgsuggest/suggestors/dotnet"
+	"go.jetpack.io/devbox/internal/pkgsuggest/suggestors/golang"
 	"go.jetpack.io/devbox/internal/pkgsuggest/suggestors/haskell"
+	"go.jetpack.io/devbox/internal/pkgsuggest/suggestors/java"
 	"go.jetpack.io/devbox/internal/pkgsuggest/suggestors/javascript"
 	"go.jetpack.io/devbox/internal/pkgsuggest/suggestors/rust"
 	"go.jetpack.io/devbox/internal/pkgsuggest/suggestors/zig"
@@ -11,7 +13,9 @@ import (
 
 var SUGGESTORS = []suggestors.Suggestor{
 	&dotnet.Suggestor{},
+	&golang.Suggestor{},
 	&haskell.Suggestor{},
+	&java.Suggestor{},
 	&javascript.Suggestor{},
 	&rust.Suggestor{},
 	&zig.Suggestor{},
