+++ internal/pkgsuggest/suggestors/golang/golang.go
